Don't use any more output of yamllint
This time I checked against yamllint.com, but I did nt use the output of this tool as I suspect yamllint scrambles somehow the file.
I took advantage to complete the constraints and I try to fix the "query", because as the query of the  political_board.yml does not work, I believe I have to change something in the query syntax as well. If Indicator_board.yml query syntax is ok, Ill fix afterwards the political_board.yml query syntax. 
Thank you